eval(session): move --save-as projection to tournament_save.rs
Pure move, no semantic change. The projection (to_saveable_config +
path rebasing) is the inverse of tournament_resolve::resolve(), but it
lived in the executor module — a maintainer changing precedence or the
TOML schema had the forward direction in one file and the backward
direction in another. tournament_save.rs names the pairing in its
module doc; tournament_run.rs is now the executor its header claims.

Not merged into tournament_resolve.rs: that file is already large, and
resolve_loaded is deliberately fs-pure while the projection
canonicalizes paths.
